New IPython Notebook
This notebook is a program that scrapes the locator.treace.net website for surgeons based on 75 zipcodes of major US metropolitan areas. The surgeons are then categorized under what procedure they perform (MIS, Lapiplasty, Both, MIS Only, Lapiplasty Only) and counted, with the results being exported as a new file in excel. The excel file contains a table with the number of surgeons under each category corresponding to each zipcode.